Further source file restructuring
