+++ connections/canconnectionmodel.cpp
