+++ connections/canconnectionmodel.h
