+++ connections/connectionwindow.cpp
