+++ connections/connectionwindow.h
