+++ mainwindow.h
@@ -8,7 +8,7 @@
 #include "canframemodel.h"
 #include "can_structs.h"
 #include "framefileio.h"
-#include "graphingwindow.h"
+#include "re/graphingwindow.h"
 #include "re/frameinfowindow.h"
 #include "frameplaybackwindow.h"
 #include "re/flowviewwindow.h"
@@ -19,7 +19,7 @@
 #include "mainsettingsdialog.h"
 #include "firmwareuploaderwindow.h"
 #include "re/discretestatewindow.h"
-#include "connectionwindow.h"
+#include "connections/connectionwindow.h"
 #include "scriptingwindow.h"
 #include "re/rangestatewindow.h"
 #include "dbc/dbcloadsavewindow.h"
