+++ re/graphingwindow.cpp
