+++ re/graphingwindow.h
