+++ re/newgraphdialog.cpp
