+++ re/newgraphdialog.h
